User authorization backend

+++ backend/routes/posts.js
@@ -39,7 +39,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost =>{
     res.status(201).json({
@@ -70,10 +71,14 @@ router.put(
       content: req.body.content,
       imagePath: imagePath
     });
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-      res.status(200).json({
-        message: 'Upate Successful!'
-      });
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+      if(result.nModified > 0){
+        res.status(200).json({message: "Update successful"});
+      }else{
+        res.status(401).json({
+          message: 'Not Authorized'
+        });
+      }
     });
 });
 
@@ -118,9 +123,14 @@ router.delete(
   "/:id",
   checkAuth,
   (req,res,next)=>{
-  Post.deleteOne({_id: req.params.id}).then(result=>{
-    console.log(result);
-    res.status(200).json({message: "Post deleted"});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result=>{
+    if(result.n > 0){
+      res.status(200).json({message: "Post Deleted!"});
+    }else{
+      res.status(401).json({
+        message: 'Not Authorized'
+      });
+    }
   });
 });
